feat: Troca valores monetarios de double para BigDecimal

+++ src/main/java/br/com/itau/geradornotafiscal/domain/aliquota/AliquotaStrategy.java
@@ -2,9 +2,11 @@ package br.com.itau.geradornotafiscal.domain.aliquota;
 
 import br.com.itau.geradornotafiscal.model.Pedido;
 
+import java.math.BigDecimal;
+
 public interface AliquotaStrategy {
 
     boolean aplica(Pedido pedido);
 
-    double aliquota(Pedido pedido);
+    BigDecimal aliquota(Pedido pedido);
 }
